Modifications

+++ helpers/tools.py
@@ -8,7 +8,7 @@ import shlex
 import asyncio
 from helpers.logger import logger
 from typing import Tuple
-
+import time
 
 async def execute(cmnd: str) -> Tuple[str, str, int, int]:
     try:
@@ -28,13 +28,24 @@ async def execute(cmnd: str) -> Tuple[str, str, int, int]:
         return ("", "", 1, 0)
 
 async def clean_up(input1, input2=None, filename=None):
-    try:
-        os.remove(input1)
-    except Exception as e:
-        logger.error(f"Error while deleting file {filename}: {e}")
-        pass
-    try:
-        os.remove(input2)
-    except Exception as e:
-        logger.error(f"Error while deleting file{filename}: {e}")
-        pass        
+    retries = 3
+    while retries > 0:
+        logger.info(f"retrying to delete file {filename} in input1: {retries}")
+        try:
+            if input1 and os.path.exists(input1):
+                os.remove(input1)
+                break
+            #os.remove(input1)
+        except Exception as e:
+            logger.error(f"Error while deleting file {filename} in input1: {e}")
+            retries -= 1
+            if retries == 0:
+                logger.error(f"Failed to delete file {filename} in input1")
+                break
+            time.sleep(1)
+        try:
+            if input2 and os.path.exists(input2):
+                os.remove(input2)
+        except Exception as e:
+            logger.error(f"Error while deleting file{filename} in input2: {e}")
+            pass
